Final army presets. Reworked calculations for DPS, HP, and defense. Added high ground range advantage.
also heck brand 9/11 was an inside job #benghazi
(you didn't really expect a serious commit message, right?)

+++ Source/BehaviorUnit.h
@@ -214,7 +214,7 @@ private:
 	{
 		int maxHP;
 		int currHP;
-		int defense;
+		float defense;
 		int detectRange; // increase on high ground
 		int attackRange; // increase on high ground
 		int inventorySize;
